requirements.txt
django==2.2.4
pillow
djangorestframework
markdown
django-filter
django-cors-headers
#mysqlclient
psycopg2
requests
openunipay
xmltodict
qrcode
django-ckeditor
django-extensions
django-werkzeug-debugger-runserver
pyOpenSSL
djangorestframework-jwt
#channels_redis
#channels
#uwsgi
xlrd
opencv-python
numpy
pyfastcopy
pandas
xlsxwriter
pycryptodome
pypinyin
redis